added water events controller and endpoints

+++ backend/home-dashboard/Models/WaterEvent.cs
@@ -5,5 +5,5 @@ public class WaterEvent
     public string? Id { get; set; } = Guid.NewGuid().ToString();
     public string? PlantId { get; set; }
     public DateTime WateredAt { get; set; }
-    public int DurationSeconds { get; set; }
+    public double DurationSeconds { get; set; }
 }
